requirements.txt
numpy==1.26.3
opencv-python==4.9.0.80
pillow==10.2.0
types-Pillow==10.2.0.20240125
